move provider related codes in base llm and llm_openai
