+++ llm_training/models/__init__.py
@@ -1,3 +1,2 @@
-from . import llm_provider
 from . import llm_training_dataset
 from . import llm_training_job
